refactor(qa): extract mock retry scenarios
